Add Method to spawner proto to spawn enemies one at a time Remove enemy queue from controller Change speed attribute of states to float Decrease the speed in each difficulty state Update the score when an enemy is killed

+++ HunterGame/EnemySubclass.cs
@@ -94,7 +94,7 @@ namespace HunterGame
         }
 
         //Method for setting all the appropriate settings based on difficulty setting
-        public void setDifficultyAttribs(int speed, int killWorth, int screenPoints)
+        public void setDifficultyAttribs(float speed, int killWorth, int screenPoints)
         {
             //Generate random starting position for the enemy
             Random Ran = new Random();
